Abstract composer trigger (#13)
* Removed redundant code
* Split event handling and logic processing from the Plugin class
* Fix missing things
* Added requiring `autoload.php`
